backend/interiitbackend/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
Django==4.0.3
finnhub-python==2.4.11
fonttools==4.31.1
idna==3.3
joblib==1.1.0
kiwisolver==1.4.0
matplotlib==3.5.1
nltk==3.7
numpy==1.22.3
packaging==21.3
pickle-mixin==1.0.2
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
regex==2022.3.15
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
sqlparse==0.4.2
threadpoolctl==3.1.0
tqdm==4.63.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
